pyproject.toml
[project]
name = "momento"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "pwdlib[argon2]>=0.3.0",
    "python-multipart>=0.0.9",
    "neo4j>=5.14.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.24.0",
    "fastapi-mail>=1.4.0",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.26.0",
]
